+++ appdaemon/assets/javascript/stream.js
@@ -69,6 +69,46 @@ var ADStream = function(transport, protocol, domain, port, client_name, on_messa
     {
         this.stream = new SocketIOStream(stream_url, this.ad_on_connect, this.ad_on_message, this.ad_on_disconnect)
     }
+    else if (transport === "sockjs")
+    {
+        this.stream = new SockJSStream(stream_url, this.ad_on_connect, this.ad_on_message, this.ad_on_disconnect)
+    }
+};
+
+var SockJSStream = function(stream, on_connect, on_message, on_disconnect)
+{
+    var self = this;
+    this.on_connect = on_connect;
+    this.on_message = on_message;
+    this.on_disconnect = on_disconnect;
+
+    this.send = function(data)
+    {
+        sock.send(JSON.stringify(data));
+    };
+
+    this.sjs_on_connect = function()
+    {
+        self.on_connect()
+    };
+
+    this.sjs_on_message = function(event)
+    {
+        var data = JSON.parse(event.data);
+        self.on_message(data)
+    };
+
+    this.sjs_on_disconnect = function()
+    {
+        self.on_disconnect()
+    };
+
+    var sock = new SockJS(stream);
+
+    sock.onopen = this.sjs_on_connect;
+    sock.onmessage = this.sjs_on_message;
+    sock.onclose = this.sjs_on_disconnect;
+
 };
 
 var SocketIOStream = function(stream, on_connect, on_message, on_disconnect)